https://online.sbis.ru/opendoc.html?guid=ea4f2115-6ae7-410b-aaec-a81010122025 В слайдере логика расчетов перенесенена в Utils
