+++ Controls/_slider/sliderTemplate.wml
@@ -17,7 +17,9 @@
                     name="{{point.name}}"
                     style="left:{{point.position}}%;">
                   <ws:if data="{{point.name=='tooltip' && _options.tooltipVisible && !_options.readOnly && (_tooltipValue || _tooltipValue==0)}}">
-                     <div class="controls-Slider__tooltipValue controls-Slider__tooltipValue_theme-{{_options.theme}} controls-Slider__point{{_options.size=='m' ? '_big' : '_small'}} controls-Slider__point{{_options.size=='m' ? '_big' : '_small'}}_tooltipValue_theme-{{_options.theme}}">{{_tooltipValue}}</div>
+                     <div class="controls-Slider__tooltipValue controls-Slider__tooltipValue_theme-{{_options.theme}} controls-Slider__point{{_options.size=='m' ? '_big' : '_small'}} controls-Slider__point{{_options.size=='m' ? '_big' : '_small'}}_tooltipValue_theme-{{_options.theme}}">
+                        {{_tooltipValue}}
+                     </div>
                   </ws:if>
                </div>
             </ws:if>
@@ -42,7 +44,7 @@
                   <ws:for data="key, interval in _intervals">
                      <div
                              class="controls-Slider__event"
-                             style="left:{{interval.left}}px; width:{{interval.width}}px; background-color:#{{interval.color}};">
+                             style="left:{{interval.left}}%; width:{{interval.width}}%; background-color:#{{interval.color}};">
                      </div>
                   </ws:for>
                </ws:if>
